# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
}

export interface Profile {
  introMessage: string;
  websiteUrl: string;
  botName: string;
  updatedAt: string;
}

export interface ProfileUpdate {
  /** @minLength 1 */
  introMessage: string;
  /** @minLength 1 */
  websiteUrl: string;
  /** @minLength 1 */
  botName: string;
}

export type BroadcastStatus = typeof BroadcastStatus[keyof typeof BroadcastStatus];


export const BroadcastStatus = {
  sent: 'sent',
  failed: 'failed',
} as const;

export interface Broadcast {
  id: number;
  message: string;
  status: BroadcastStatus;
  /** @nullable */
  errorMessage?: string | null;
  sentAt: string;
}

export interface BroadcastInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
}

export interface Stats {
  /** @nullable */
  followerCount: number | null;
  /** @nullable */
  followerCountDate: string | null;
  /** @nullable */
  targetLimit: number | null;
  /** @nullable */
  totalUsageThisMonth: number | null;
  broadcastCount: number;
}