# lib/api-zod/src/generated/types/broadcast.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BroadcastStatus } from './broadcastStatus';

export interface Broadcast {
  id: number;
  message: string;
  status: BroadcastStatus;
  /** @nullable */
  errorMessage?: string | null;
  sentAt: string;
}