# lib/api-zod/src/generated/types/broadcastInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BroadcastInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
}