# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  /** @nullable */
  followerCount: number | null;
  /** @nullable */
  followerCountDate: string | null;
  /** @nullable */
  targetLimit: number | null;
  /** @nullable */
  totalUsageThisMonth: number | null;
  broadcastCount: number;
}